GSWB-83: Modifying script to delete tar always

+++ deploy.sh
@@ -7,6 +7,8 @@ DockerVersion=1.0.0
 find . \( -name "*.csproj" -o -name "*.sln" -o -name "NuGet.config" \) -print0 \
     | tar -cvf projectfiles.tar --null -T -
 
-docker build . -t thegoodframework:$DockerVersion -t thegoodframework:latest
-
-rm projectfiles.tar
+docker build . -t registry.guildswarm.org/base-images/thegoodframework:$DockerVersion -t registry.guildswarm.org/base-images/thegoodframework:latest
+if [[ $0 == 1 || 0 ]]
+then
+    rm projectfiles.tar
+fi
